Create common module for microsoft directline API methods
